+++ src/Controller/GroupController.php
@@ -135,6 +135,12 @@ final class GroupController extends AbstractController
     )]
     public function update(#[MapEntity(id: 'id')] Group $group): Response
     {
+        if ($group->getIsSystem()) {
+            return new JsonResponse([
+                'error' => 'System groups cannot be updated.'
+            ], Response::HTTP_FORBIDDEN);
+        }
+
         if (!$this->deserializer->deserialize($group, ['update'])) {
             return $this->deserializer->respondWithError();
         }
@@ -167,6 +173,12 @@ final class GroupController extends AbstractController
     )]
     public function delete(#[MapEntity(id: 'id')] Group $group): Response
     {
+        if ($group->getIsSystem()) {
+            return new JsonResponse([
+                'error' => 'System groups cannot be deleted.'
+            ], Response::HTTP_FORBIDDEN);
+        }
+
         $this->entityManager->remove($group);
         $this->entityManager->flush();
